+++ bsdploy/fabfile_mfsbsd.py
@@ -124,6 +124,10 @@ def _bootstrap():
     print("bootstrap-autoboot-delay:", autoboot_delay)
     bootstrap_reboot = value_asbool(env.instance.config.get('bootstrap-reboot', 'true'))
     print("bootstrap-reboot:", bootstrap_reboot)
+    bootstrap_packages = env.instance.config.get('bootstrap-packages', 'python27').split()
+    if firstboot_update:
+        bootstrap_packages.append('firstboot-freebsd-update')
+    print("bootstrap-packages:", bootstrap_packages)
 
     yes = env.instance.config.get('bootstrap-yes', False)
     if not (yes or yesno("\nContinuing will destroy the existing data on the following devices:\n    %s\n\nContinue?" % ' '.join(bu.devices))):
@@ -165,9 +169,7 @@ def _bootstrap():
     bu.create_bootstrap_directories()
     bu.upload_bootstrap_files(template_context)
 
-    bootstrap_packages = ['python27']
     if firstboot_update:
-        bootstrap_packages.append('firstboot-freebsd-update')
         run('''touch /mnt/firstboot''')
         run('''sysrc -f /mnt/etc/rc.conf firstboot_freebsd_update_enable=YES''')