+++ CHANGES.rst
@@ -2,6 +2,7 @@
 ====================
 
 - [feature] allow override of ``destroygeom`` via ``bootstrap-destroygeom``.
+- [feature] allow override of packages installed during bootstrap via ``bootstrap-packages``.
 
 
 3.0.0b3 - 2019-06-09